+++ packages/ok-components/CHANGELOG.md
@@ -1,8 +1,17 @@
 # Change Log - @ni/ok-components
 
-<!-- This log was last generated on Thu, 11 Sep 2025 16:32:48 GMT and should not be manually modified. -->
+<!-- This log was last generated on Wed, 01 Oct 2025 05:25:42 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.0.3
+
+Wed, 01 Oct 2025 05:25:42 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v34.3.1
+- Bump @ni/spright-components to v6.3.1
 
 ## 0.0.2